src/test/java/edu/hw1/Task8_KnightsOnBoardTest.java
package edu.hw1;

import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
import static org.junit.jupiter.api.Assertions.assertTrue;
import static org.junit.jupiter.api.Assertions.assertFalse;
import static org.assertj.core.api.Assertions.assertThat;
import org.junit.jupiter.api.Test;

public class Task8_KnightsOnBoardTest {

    @Test
    public void knightBoardOnWhichNoKnightCanCaptureAnotherKnightShouldReturnTrue() {
        assertTrue(Task8_KnightsOnBoard.knightBoardCapture(new int[][] {
            {0, 0, 0, 1, 0, 0, 0, 0},
            {0, 0, 0, 0, 0, 0, 0, 0},
            {0, 1, 0, 0, 0, 1, 0, 0},
            {0, 0, 0, 0, 1, 0, 1, 0},
            {0, 1, 0, 0, 0, 1, 0, 0},
            {0, 0, 0, 0, 0, 0, 0, 0},
            {0, 1, 0, 0, 0, 0, 0, 1},
            {0, 0, 0, 0, 1, 0, 0, 0}}));
    }

    @Test
    public void knightBoardOnWhichThereIsKnightCanCaptureAnotherKnightShouldReturnFalse() {
        assertFalse(Task8_KnightsOnBoard.knightBoardCapture(new int[][] {
            {1, 0, 1, 0, 1, 0, 1, 0},
            {0, 1, 0, 1, 0, 1, 0, 1},
            {0, 0, 0, 0, 1, 0, 1, 0},
            {0, 0, 1, 0, 0, 1, 0, 1},
            {1, 0, 0, 0, 1, 0, 1, 0},
            {0, 0, 0, 0, 0, 1, 0, 1},
            {1, 0, 0, 0, 1, 0, 1, 0},
            {0, 0, 0, 1, 0, 1, 0, 1}}));

        assertFalse(Task8_KnightsOnBoard.knightBoardCapture(new int[][] {
            {0, 0, 0, 0, 1, 0, 0, 0},
            {0, 0, 0, 0, 0, 1, 0, 0},
            {0, 0, 0, 1, 0, 0, 0, 0},
            {1, 0, 0, 0, 0, 0, 0, 0},
            {0, 0, 0, 0, 1, 0, 0, 0},
            {0, 0, 0, 0, 0, 1, 0, 0},
            {0, 0, 0, 0, 0, 1, 0, 0},
            {1, 0, 0, 0, 0, 0, 0, 0}}));

        assertFalse(Task8_KnightsOnBoard.knightBoardCapture(new int[][] {
            {0, 0, 0, 1, 0, 0, 0, 0},
            {0, 1, 0, 0, 0, 0, 0, 0},
            {0, 1, 0, 0, 0, 1, 0, 0},
            {0, 0, 0, 0, 1, 0, 1, 0},
            {0, 1, 0, 0, 0, 1, 0, 0},
            {0, 0, 0, 0, 0, 0, 0, 0},
            {0, 1, 0, 0, 0, 0, 0, 1},
            {0, 0, 0, 0, 1, 0, 0, 0}}));
    }

    @Test
    public void invalidBoardSizeShouldThrowsException() {
        assertThatIllegalArgumentException().isThrownBy(() -> {
            Task8_KnightsOnBoard.knightBoardCapture(new int[][] {
                {0, 0, 0, 0, 1, 0, 0, 0},
                {0, 0, 0, 0, 0, 1, 0, 0},
                {0, 0, 0, 1, 0, 0, 0, 0},
                {1, 0, 0, 0, 0, 0, 0, 0}});
        });

        assertThatIllegalArgumentException().isThrownBy(() -> {
            Task8_KnightsOnBoard.knightBoardCapture(new int[][] {
                {0, 0, 0, 1},
                {0, 1, 0, 0},
                {0, 1, 0, 0},
                {0, 0, 0, 0},
                {0, 1, 0, 0},
                {0, 0, 0, 0},
                {0, 1, 0, 0},
                {0, 0, 0, 0}});
        });

        assertThatIllegalArgumentException().isThrownBy(() -> {
            Task8_KnightsOnBoard.knightBoardCapture(new int[][] {
                {0, 0, 0, 1, 0},
                {0, 1, 0, 0, 1},
                {0, 1, 0, 0, 0},
                {0, 0, 0, 0, 0},
                {0, 1, 0, 0, 0}});
        });
    }
}
